fix(backend): resolve Swagger auth, Dapper type mismatches, and EF Core query warnings
- Add Bearer JWT security scheme to OpenAPI spec (Authorize button in Swagger UI)
- Guard against missing JWT secret in non-development environments (fail-fast)
- Fix COUNT(*) int→long type mismatch in all Dapper analytics DTOs
  (GenreCountDto, DirectorCountDto, ActorCountDto, DecadeCountDto,
   MonthlyActivityDto, MonthlyWatchesDto, WeeklyWatchesDto, DirectorLeaderboardDto)
- Add OrderBy(m => m.Id) before Take() in EnrichPendingMoviesCommandHandler
- Configure SplitQuery globally via UseNpgsql to eliminate cartesian explosion warning
- Suppress EF Core SQL command and query log noise in appsettings.Development.json

+++ backend/Frametric.Application/DTOs/Analytics/ActorCountDto.cs
@@ -1,3 +1,3 @@
 namespace Frametric.Application.DTOs.Analytics;
 
-public record ActorCountDto(string ActorName, int Count, decimal AverageRating);
+public record ActorCountDto(string ActorName, long Count, decimal AverageRating);
