+++ backend/Frametric.Application/DTOs/Analytics/GenreCountDto.cs
@@ -1,3 +1,3 @@
 namespace Frametric.Application.DTOs.Analytics;
 
-public record GenreCountDto(string GenreName, int Count);
+public record GenreCountDto(string GenreName, long Count);
